Fix warnings about deprecated behaviour of QByteRef
Make sure that the dst QByteArray is big enough and filled up with
known content before copying the rle'd src into it. Then when done
resize to the final size.

This prevents the following from being printed a gazillion times
when loading an image with a layerstyle that uses a pattern:

Using QByteRef with an index pointing outside the valid range of a QByteArray. The corresponding behavior is deprecated, and will be changed in a future version of Qt.

BUG:419289