Fix MyPaint default brush to be a brush that paints
Before this commit, if the user created a new MyPaint brush
using the + button in the left panel of the Brush Editor,
it would by default create a brush that doesn't have an ability
to make marks on the canvas (the dab-per-second and dabs-per-radius
are both 0, the opacity is 0, the diameter is very small as well).

This commit adds a default preset for MyPaint that can paint,
and changes the code to make use of that preset.

Note: right now Krita ignores MyPaint's default settings for the brush.
Since all other paintop engines have their default presets in the
.qrc file, it was the easiest way to make MyPaint brush engine work as
well.

BUG:434309