Fix extra call to tool->popupWidget() if tool->popupActionsMenu() returned something
By a common convention, the newer version function should not be
called when the other version has returned something.

It is quite disputable, what version should be called first, but
that is mostly a "theoretical" question in our case (since it is
not a public API).

(Squashed with some other miscellaneous cleanup changes. - Emmet)