Remove JPG from the list of supported clipboard image formats
Mint randomly puts a JPG in its set of images on the clipboard
in its screenshot tool. And if it's the first image in the set,
Krita will try to use it, which gives pastes with JPG artefacts.

Gimp doesn't accept JPG clips, which makes people think it's a bug
in Krita, while it's just Mint being dumb. Anyway, there's no
good reason to accept JPG clips, and I'm not sure why we ever did.

BUG:431310