Improve default style of Button.qml
Make the background and highlight transparent by default,
so they can be better integrated on specific background like in the dialogs.

Also make the icon semi-transparent on pressed when the highlight is transparent,
and reduce by half the duration of the opacity animation on it to make it look
more responsive.