Use ButtonSquared also for undo-redo touchstrip buttons
Use ButtonSquared also for undo-redo touchstrip buttons.

It is more consistent to use same button style on those too.

Now only open/save/save-as buttons don't use it on purpose,
as they look better with the specific circled icons,
and as they are a sort of different category it can make sense.