Fix font resizing on qml for the touch docker
Fix font resizing on qml for the touch docker.

Do not use font settings from the theme,
as this does not make the font size adapt to the window size dynamically.

Better directly set the font options in the qml,
using the Constants values for the sizes, as this ensures that text size
is always dynamically adapted.