Fix mirroring for MyPaint brushes
Before this commit, MyPaint brushes wouldn't work
with mirroring at all (the brush wouldn't paint
on the other side of the axis). This commit introduces
mirroring to MyPaint brushes.

There is still an issue when the layer you paint on has
different contents on both sides of the axis, since
the other side would get some "halos". But it's the same
issue as with Color Smudge and other engines.

BUG:434307