Reduce the amount of memory needed for MyPaint masks
Before this commit, MyPaintSurface would write a mask
to a paint device at the place where dab was located.
That's not memory efficient, because after some time
of painting, a huge area might be needed.
This commit ensures that the mask always starts at (0,0),
which means that the area is only as big as the biggest
dab.