Add a note about potential UB in KisFixedPoint class
Bitwise shift operators for signed (and negative) integers
became officially supported only in C++20. Though it looks
like our GCC implementation supports them fine even in earlier
versions of the standard.

We have a unittest that tests this behavior, so it should be
fine to use this (taking into account that this code is used
in Krita since 2012 ;) )