Fix crash on exit by setting styleSheet only on QTabBar
Before this commit, Krita would crash on exit on Qt 5.12.8
with segfault in the destructor for the Python plugin
widgets. It turned out that this crash was introduced in
8a7bae681a1f992, and the reason turned out to be
a method setStyleSheet() with non-empty string.
Setting empty stylesheet later (on closing) didn't make a difference.
However setting the stylesheet on just the widget
that needs it (QTabBar that keeps the document tabs)
fixes/prevents the crash.

Alternative fix would be to find the widget for the Python plugin
and explicitely set empty stylesheet after every non-empty
stylesheet set for the whole application. It wasn't tested,
but in case more widgets need non-empty stylesheet, it sounds
like it should work.

BUG:427137
BUG:420612