Fix 2 Point Perspective not respecting Snap Single
Before this commit, since 2 Point Perspective was in fact one assistant
(even though it was composited out of three and had three options
of the line inside), the Snap Single feature wouldn't work on it
(the user would see that the line it draws snaps between different
vanishing points). This commit ensures that 2PP remembers the last used
vanishing point and sticks to it later if the Snap Single feature
is turned on.

Test case:
- make a 2PP (you can add more assistants too)
- select brush tool, make sure Snap to Assistants and Snap Single
are both checked
- start a line, then come back to the point you started from,
but in a big distance from the line - you should see how
2PP jumps between different vanishing points, making your line
not straight anymore but composited out of multiple short straight
lines.