fix "make test" on macOS
The number of failed tests on macOS is significantly higher as
many tests require to be installed to be run, this does not mean
the test is broken, but that we haven't come up with a good way to make sure Qt finds the bundle path for test apps running from the build directory. Such tests can be run if copied to the install dir and run manually.