Bugfix: Artifact with more than one active assitant
Problem: With more than one assitants active on canvas, often on devices
with high resolution the Pixmap cache would fill up immediately. This
would internally trigger glDeleteTexture right after the first pixmap is
rendered and when turn for second pixmap is due, it would get assigned
the same name as the previously generated (which is a correct behavior!
but it only seems to happen with some drivers). With same name as
previous one Qt doesn't render it (see
QOpenGL2PaintEngineExPrivate::updateTexture).

Workaround: Because of this name clash, we workaround this Qt bug by
increasing the cache size, so the new textures don't immediately delete
the previous ones and take their names. Ideally, this should be fixed
inside Qt.

Devices with this problem:
+ Android
+ ChromeOS (mesa)
+ Windows (ANGLE)

BUG:401940