Fix KisAlgebra2D::intersectLineRect()
Before this commit, function intersectLineRect
was both buggy and confusing to use.
This commit fixes this behaviour to be more
like what's expected, and extends it to be used
in more cases more easily.

Bugs/problems before this commit:
- if you put line(A, B) to the function,
the returned line will often be (B', A') instead
of (A', B')
- it used bottom() etc. to determine the edges,
which because it was QRect, not QRectF, was 1 unit
smaller than the actual value. That made the resulting
line incorrect.
- it always extended the line to the edges of the rectangle.
While it can sometimes be the desired behaviour,
it's not explained in the documentation of the function.

After this commit, the first two problems are fixed,
and the last one is resolved so that now the developer
can specify if they want to extend the line to the edges
of the rectangle or not.

Changes in the assistants and mirroring tool preview
drawing code are necessary because of the fix for the
issue #2, which reversed the order of the points in
the returned line.

---
Note: there is one more issue not addressed in this commit.
The API of the function states that if the line doesn't
intersects the rectangle, `false` is returned and the line
is not changed. This is not necessary convenient, since now
for example in case of "just draw the line inside the viewport"
the developer cannot just draw the line, they need to first
check the returned value and then decide whether to draw it or not.
Note that presumably lots of assistants don't check the return value
and draw the lines unnecessarily outside of the viewport.