add macos script for running all compiled tests
This is a convenience script to run all compiled tests.app from
the current dir recursively. (you can run all tests found or just
a set of  argument given tests)

On macos some tests require to be run onn the install dir for Qt to find the bundle path, this tests are marked as broken to allow 
ctest to run with no errors.

The script will use BUILDROOT to se the install directory and to
check if the directory you are running the tests is child of
to the BUILDROOT directory. (this can be overriden with args
--install_dir=<path> and --work_dir=<path>) From the run
directory the script will find any .app bundle, install it, and
run it, reporting if the test failed or passed.

The output is very similar to ctest, however this does not filter
real broken tests than bundle_path broken tests.