Fix RGBA brushes created in Krita 4.4 becoming darker in Krita 5.0
In Krita 4.x releases there was a bug that caused lightness
adjustments to be applied to the brush **twice**. It happened
due to the fact that copy-ctor called brushTipImage() virtual
method instead of just copying the image itself.

In Krita 5 we should open these brushes in somehwat the same way.
The problem is that we cannot convert the numbers precisely, because
after applying a piecewice-linear function twice we get a
quadratic function. So we fall-back to a blunt parameters scaling,
which gives result that is just "good enough".

BUG:436257

# Conflicts:
#	libs/brush/KisColorfulBrush.cpp
#	libs/brush/kis_predefined_brush_factory.cpp