KisAnimTimeline: Fixed inconsistent scrollTo behavior.
When switching active layer there is a brief window of time where
two unpinned layers simultaneously exist on the timeline, causing
indexing issues that result in improper scrollTo behavior for the
last layer.

This commit adds a workaround by delaying the scrollTo for 1/60
of a second to give the timeline model enough time to have correct
indexing.

BUG:437029