Fix disabling preview of 2pp assistant
Before this commit, 2pp assistant preview was impossible
to turn off in the View menu. This commit ensures that if
Show Assistant Preview is turned off, the grid disappears.
Note: the behaviour is slightly different from Vanishing
Point Assistant, because it does show the preview if
"Show Painting Assistant" is off and the "Show Preview" is on.
However this seemed more correct, and is consistent with
other assistants like Parallel Ruler.

BUG:437415