Add symbolic visual representation for vanishing points on editing
Before this commit, when placing points when making the 2pp,
it wouldn't be obvious what the points do until you place the second
one. For 2pp assistant it's not the worst, but it makes a difference
for the limited area/local 2pp assistant (in the works).
This commit will change the behaviour on both (when 2pp assistant
will be merged too).