Fix xml created by 2pp assistant
Before this commit, the xml created was - surprisingly -
correct xml-wise, but it was incorrect semantic-wise.
It would put useVertical, handles and sidehandles
elements inside the gridDensity element.
It didn't cause any user-visible issue I could find,
but when implementing local 2pp assistant, I had issues
caused by this, so I expect that this might have
unexpected undesired effects down the line.

After this commit, the xml is correct semantically,
and all elements: gridDensity, useVertical,
handles and sidehandles are children of the
assistant element.