KisAnimCurves: Added double click fast selection methods to curves docker as well.
I'm not in love with the way select + drag operations work currently. There are cases
where click and dragging to close to the handles will cause a deselection instead of
a click drag. We might need to store intention of deselection across to release event,
and check if any "dragging" occured along the way to prevent accidental deselection.

I'll work on it more tomorrow.

BUG:437631