Set lastPoint in TwoPointAssistant to -1 in new stroke
Before this commit, the m_lastUsedPoint was not changed to -1
on every new stroke, which could potentially cause issues
with aligning on subsequent strokes.
This commit always ensures it starts as -1 (so it's treated as
no information - this is a per-stroke variable so it needs
to be reset every time).

This commit also contains a little fix in documentation
with related code.