Add the angle selector to the pop up palette
This adds a mechanism so that paintop properties can have a subtype associated
with them (for example "angle", "color", etc.). Right now it is used to
indicate if a int/double property is an angle. Then that is information is used
to dynamically use a slider or an angle selector.