Fix losing off-canvas data when doing perspective trasnform
Basically, off-canvas cropping is needed only for the transform
mask. The transform tool should transform the entire image all the
time.

The only danger of this change is that perspective transform can
easily grow to infinity. But I didn't manage to reproduce this growth
with the transform tool. I guess it is only possible with the transform
mask (when the source layer is moved).

BUG:418406