Popup Palette: Improves UX related to bottom bar toggle
Moves the clear colors button outside the bottom bar since that is
something related to the palette and would not be accessible with the
bottom bar collapsed.

Moves all the buttons closer to the palette so it looks nicer when it's
closed.

Adds a slight margin beneath the popup palette to match the margin to
the right for the brush hud, otherwise the buttons around the palette
looked off center.

Moves the reset zoom button to the right so that the zoom bar is not off
center in the bottom bar.

GUI: