Popup Palette: Add setting to hide/show rotation track
This also adds a tiny margin around the presets ring, otherwise when the
rotation track is hidden, the slots pop out of the palette.

The inner rotation track is also thicker so that it's width can be taken
into account when laying out the presets and the margin is pixel perfect.