Fix saving palettes in the folder of appimage
Before this commit, the palette was saved unnecessarily,
because the saving is done by updateResource() as well.
This could cause the issue of new versions of the palette
in the folder of the appimage instead of in the resource
folder.
This commit omits the manual saving and relies on the resource
system to do it.