Fix subfolders with bundles
Before this commit, the BundleStorage would report
the resources inside it incorrectly, which is,
without the subfolder. This commit ensures that the
BundleStorage adds the subfolder to the resource path.
Note: it only works for the resources in the bundle
itself; for the resources in the _modified/,
I think they are just put in one folder all together,
without any subfolders.