Ask for an icon when overwriting a brush preset
Before this commit, the user had no way to:
1) cancel the overwrite,
2) change the icon of the brush preset without loading
the .kpp file into Krita (which is much more troublesome
than this little useful dialog with bunch of vector shapes).

This commit enables the icon dialog also on overwrite which
fixes both of those issues.
Note: this doesn't include any user string changes, so it can
be pushed on master during string freeze without an issue.