Make Brush Presets the default res type in Resource Manager
Before this commit, the default resource type was Brush Tips
instead of Brush Presets, and since the default storage is
Local Resources, it caused the list to only show
the triangle brush tip that is used when the actual brush tip
cannot be found. This was a bit ugly/unwelcoming to users
(since they mostly expect their "brushes" (brush presets)
to be visible and they want to organize those).
This commit changes it so the default resource type on opening
is Brush Presets.