Allow writing 0 bytes to KoQuaZipStore
Before this commit, if you tried to write 0 bytes to KoQuaZipStore,
it would return "false" as if there was an error, even though
the number of written bytes and the number of bytes in the cache
were equal. In fact, the previous "error-checking" wouldn't
catch an error, because it only checks it for 0, while write()
returns -1 on error.

This commit ensures that write() returns false iff the number of
written bytes is different from the number of bytes the user
expected to be written.