Fix saving new gamut masks
Before this commit, gamut masks docker would use
KoResourceServer to manage resources, and always add the gamut
mask as a new resource, even if it already existed.
This commit makes it use KisResourceModel instead
(the newer resource API) and only updates the resource
instead of adding it again.

Right now creating a new mask should work as long as the name
doesn't change (which is possible when editing the newly created
mask).

CCBUG:438033