Fix duplicate mask in Gamut Masks docker
Before this commit, duplicating a mask would just edit the old one.
That's because the resource id wasn't reset to -1 after cloning.
This commit ensures the resource id is reset, so later
it won't be able to find the resource in the resource model
and it will add it to Krita as a new resource.

BUG:438033