Fix setting 'modified' state to readonly documents
Before this commit, the readonly documents would accept
first setModified(true) call, but then ignore
setModified(false) call because "it's read-only so it cannot
be modified".
That's because it was checking d->modified (current state)
instead of mod (the desired state).
That caused gamut masks (which are marked read-only) to ignore
setModified(false) call and ask the user to save the document.