Fix updates when transforming a group layer with a tranparency mask
Before the patch, the updates were issues by individual transformation
commands, that leads to weird results when the group is under an effect
of a transparency mask. So now the updates are gathered in buckets and
are emitted all at once.

The patch is still a bit clumsy, because undo/redo do not call
image->disableDirtyRequests()/enableDirtyRequests(), whereas the
main transformation action does. Though it shouldn't cause too many
issues, just a bit more updates than needed. It will still be better
than before.

BUG:431756