Fix saving layer styles
Before this commit, saving layer styles wasn't possible
because it didn't have md5. This commit ensures that
the layer styles gets an md5 before saving.
Note that md5 is calculated using the QByteArray from
saving to the QBuffer using the layer styles serializer.
It might be a better idea to just make an md5 out of
uuid.