Fix layer styles being invalid after getting loaded
Before this commit, the layer style loaded from the document
couldn't be added to the Memory Storage because it was invalid.
This commit fixes the validity of the layer style
after being loaded.

(Note that the layer style still cannot be saved to the Memory
Storage due to other reasons).