ki18n: fix loading our translations from Windows
This commit introduces a patch to k18n, suggested by Alvin Wong, to make
ki18n behave properly with our patched copy of gettext. The latter
contains a patch of vcpkg to make it behave with UTF-8 paths on Windows,
but this wasn't reflected on ki18n's initialization code.

The patch works by inserting an additional define in libintl.h that
flags the presence of vcpkg's UTF-8 patch. While at it, we update the
libgnuintl.win32.h template file to match the underlying gettext
version.

BUG: 438421

CCMAIL: kimageshop@kde.org