Popup Palette: Fix rotation handle not selectable after mirroring canvas
Before this patch, when the user mirrored the canvas, while the paint
event was correctly painting the handle since it was getting the canvas
angle directly, the cached rect that was used to check for it's hover
state was not being updated, making the handle unselectable.

This patch makes it so that the cached rect
m_canvasRotationIndicatorRect is updated correctly and used to paint the
handle.

It also removes the absolute argument of rotationIndicatorRect (it is
now always absolute) since it is not really needed anymore with the way
the code has been organized.