Fix MemoryStorage for resource that cannot saveToDevice()
Before this commit, there was no way to add a layer style to the
MemoryStorage, because the layer style cannot be saved into
bytes.
This commit fixes this behaviour except for one function,
which is loadVersionedResources().
It shouldn't really be used in Krita, at least for now,
so this commit adds a Safe Assert there instead.

For all resources that can be saved into the device,
it shouldn't change the way they are stored.