Avoid saving gamut masks prematurely
Before this commit, a gamut mask would always be saved
during creation, and then after the user edited it to be
what they wanted it to be. It meant two issues:
- the first usable version of the gamut mask was actually
the second version in the versioned_resources,
- if the user cancelled the creation and then tried to create
a gamut mask under the same name, it would fail

This commit fixes it by ensuring that saving only happens
when the user presses "Save" button.

BUG:438033