Change debug to not flood the console
The "No MD5 for..." warning always happen and will happen with layer
styles, because before overwriting the md5, we ask for the md5 to see
if it's empty. It will be empty unless set, because layer styles cannot
save to device.
No need to have it in the general logs; and if someone is interested in
md5 specifically, then can enable resources debug or change this line.