Fix "filter by tag" when checked on startup
Before this commit, if the user opened Krita and "filter by tag"
was enabled and the user tried to search for some brushes,
Krita would ignore the checked "filter by tag" option
and show all brushes with all tags. Only after checking
and rechecking the bool it would work.
This commit fixes that behaviour by only showing the relevant
brushes.

BUG:438818