Fix patterns not added to predefined from Custom Pattern
Before this commit, if the user clicked "Add to Predefined
Patterns" in the Custom Pattern widget (in the pattern
popup), the pattern wasn't added to the predefined patterns.
This commit fixes that.

Note that the pattern chooser doesn't show the new pattern.
The Resource Manager shows it though, and the pattern is shown
on the next opening of Krita.