Fix showing metadata in Bundle Manager
Before this commit, the bundle manager didn't show
the website of the bundle and showed dates wrong.
Note that dates are, at least for now, stored in
string form.
This commit fixes it by showing dates in raw strings
(exactly like they are stored) and by updating the
website label.