Fix gradients tooltip in Resource Manager
Before this commit, gradient tooltip wouldn't work properly
in the Resource Manager. It would show up very long
and (probably) 1px wide, which is why it looked as if there
was nothing there. This commit fixed that by using the same
values as the gradients chooser would use.

This commit also adds a little fix for the presets chooser
to always disable showing checkers when showing patterns.