Fix unrecognized metadata in older bundles
In older bundles, the metadata was written this way:
<meta:meta-userdefined meta:name="license" meta:value="CC-BY-SA"/>
and in the new bundles, there are two kinds of metadata
and the new kind of writing the same metadata is:
<meta:license>CC-BY-SA</meta:license>
Before this commit, Krita 5 would only read the new way of
defining the email, license and website information.
This commit fixes it by ensuring the old way of writing metadata
is still supported.