Find default tag for popup palette
Before this commit, the code would use an old method
to get the tag it wanted, and it wouldn't have a way
to get the default tag (My Favourities) for non-English
systems. Because user strings cannot be changed for now,
this commit includes a hack that tries to find first tag
with star in in, assuming it's the Favourities tag.

CCBUG:438470