Use correct context for translated resource names
Before this commit, CacheDb wouldn't use any context,
which meant that translated resource names wouldn't be used
(because the translated names had a context).
This commit provides that context so a lot more resource
names are translated.

Note: I can't say for sure if the context is correct for
all resources. If you find a resource that should have been
translated in the database but isn't, please reopen the
bug report.

CCBUG:438470