Remove unnecessary hack in Favourite Resource Manager
Before this commit, Favourite Resource Manager tried to find
"My Favourites" tag in the tag model by looking for the star
symbol. This is unnecessary because earlier in the code
it tries to find the correct tag by url.
Urls are not translatable and cannot be changed, so it should be
enough to just look for the url, no matter the language.

CCBUG:438470