update svgz icon to use svg gradients instead of embedded png
update svgz icon to use svg gradients instsead of embedded png

This improves the rendering of the svg icon in dolphin, plasma, and other places
supporting only TinySVG, and it's much lighter in size.

I've checked a visual diff (using difference blending mode) between the 2 versions,
and they look perfectly the same, so re-rendering the various png versions
is not even needed.