Implement loading and saving of embedded resources in a  preset
This is a generic implementation, so once we add gradients other
than the current one, for instance, it'll just work (tm).

Note that we also save the md5sum, so on loading we ignore the
embedded resource if there's a resource of the right type with
this md5sum (because they are obviously identical in that case).