Add resources from bundles' names to translations
Before this commit, resources from the bundles
didn't have translations. This commit ensures that the resources
in the bundles can have their names translated.

To make sure the resources of the same type have the same code
for reading both from the folder and from the bundle,
functions readGeneric() and readAllLinesGeneric() were written
to allow just calling the function that will take care of
the specific syntax of reading from the file.
Note that readGeneric() takes a different hash for the zip
handle than readGenericAllLines (explanation below).

Note:
$ziph->read() didn't work correctly (always returning empty string)
so I used a hack with reading the whole file using getline().
Possibly I was using read() incorrectly; but this method works.
If anyone in the future wants to improve this to use read()
directly, without the hack, feel free to.