TIFF: add support for reading Photoshop layers and metadata
This commit builds upon the previous work in order to enable loading
the PSD-like layer data and metadata from the TIFFTAG_IMAGESOURCEDATA
and TIFFTAG_PHOTOSHOP fields.

Of special relevance are the following:

TIFFTAG_IMAGESOURCEDATA is a sequence of Layr/Lr16/Lr32 layer block,
followed by LMsk (which is Adobe's take on a global mask info block),
followed by Patt and Anot. As these are a byte-by-byte equivalent
of the PSD's additional layer info blocks, for TIFF I've added a logic
path that takes advantage of Lr16/32 and handles them all as additional
logic blocks safely. It must also be noted that these are globally
aligned to 4-byte boundaries, which do *not* apply to internal blocks
(see Layr), and they are also in *the endianness of the container* (this
is the reason why I added endianness handling in previous commits).

TIFFTAG_PHOTOSHOP is just a concatenation of resource blocks (as opposed
to a full resource section), but unlike the previous, it's in big endian
like its PSD sibling.

BUG: 338470