Switch to using a hex-encoded QString md5sum in KoResource
Converting to and from hex-encoding when storing the md5sum in
configuration files or the database can easily lead to bugs, where
a conversion step is forgotten, or where the wrong data is fed to
the QByteArray::fromHex, which doesn't do any error checking.

This commit centralized the hex-encoding in the KoMD5Generator,
and the rest of Krita uses nothing but that hex-encoded string.