Added collapsible button group to KisAnimCurvesDocker.
Allows a set of buttons / actions to contract down to a single
click and hold button.

If the widget has enough space, it will show all buttons in the group.

If it doesn't have enough space, it will collapse to  a single button.
The single button that's shown will be whichever button in the group that
was last used. Click and holding the collapsed button will open a menu
where other buttons in the group are accessible.

Buttons in a KisCollapsibleButtonGroup should ideally be logically related.

CCBUG:439679