Fix bug on GNOME system where collapsible buttons were not rendering.
Since Eoin tested this on a KDE machine, he didn't notice that the
action icons weren't visible on the dropdown. This fixes that,
and adds some already localized text elements by using the tooltip.

Once we're out of string freeze, we should try making a set of unique
strings here...

CCBUG:439679