Re-implement the "broken" overlay icon for the preset chooser.
We now check on loading the preset whether its dependent resources
are available, if it's an old-style preset that doesn't embed them.

That sets a broken flag in the metadata, which the we can read in
the preset chooser. Note: only on creating the database or adding
a new preset, this is checked. We don't check whether the user
actually removes a dependent resource after creating the database.

In the future, we could check this status on loading a preset and
update the database accordingly.