Fixed Regression: Allow odd sizes for animation export
Initially, a patch was done that rounded animation export to even sizes
to prevent mp4/mkv crashes. Since then, mustHaveEvenDimensions has been
implemented specifically for mp4/mkv to round the dimensions.

This patch removes the old duplicate code, and allows other formats
that are not mp4/mkv to export to odd sizes.