Make default configuration of the Halftone filter correspond to global defaults
It basically switches default interpolation from 'sine' to 'linear'.

The bug in question is not really correct. The default halftone shape
is **not** diamond. Instead, it is round with sine interpolation. But
when the user switches to any other shape and returns back to round,
then the interpolation automatically resets into linear. That is extremely
confusing.

This patch just ensures that linear mode is used as default all
the time.

BUG:430281