Remove signal compressor on KisTagSelectionWidget
The signal compressor is no longer needed in that class.
It also caused trouble down the line when a new dialog
was added in one of the relevant classes.
This commit removes the signal compressor completely.

If there will be any issues, using a signal/slot connection
instead of calling the slot directly is an option.