Add allowOverwrite option when adding a tag in KisTagModel
Before this commit, KisTagModel would always just undelete
the tag if it was asked to add a tag with the same url
as an existing tag.
This commit changes this behaviour to overwrite the tag
(which means the new Name and Comment is used and all
assigned resources are unassigned) if "allowOverwrite"
is set to true, or just return false if "allowOverwrite"
is set to false.