Ask user on adding a tag with the same url in TagChooser
Before this commit, if the user tried to create a new tag
with the same url as an existing tag,
it would undelete the deleted tag, which means the assigned
resources would still be assigned.
This commit instead asks the user what they want to achieve,
and there are three options:
* Replace (overwrite) - this will reset assigned resources
* Undelete - this will cancel adding the tag and just
undelete the existing, deleted tag with all the assigned
resources it had before
* Cancel - no operation

This will be for both the tag tool button popup and
for the right-click in the resource chooser
(KisResourceItemContextMenu) because it uses the same
code.

BUG:439423

CCMAIL:kde-i18n-doc@kde.org