Use KisGlobalResourceInterface::resource in KoResourceServer
KoResourceServer is the legacy api we use to access resources
because there's much too much places for direct porting to
KisGlobalResourceInterface or KisResourceModel.

However, KoResourceServer is used for two purposes: working with
resources from the ui, like setting resources inactive, and
retrieving resources from configurations and settings.

For the first, we need resources filtered by active/inactive,
for the other, we need to be able to retrieve inactive resources.