Fix an assert in KisStrokesQueue when cancelling LoDN stroke too quickly
When the user cancells the stroke with Esc key and starts a new stroke
with instant preview too quickly, the queue may enter a race condition,
when the current lod range is not estimated correctly. To overcome this
issue, we now add a fake legacy stroke into the end of the queue to
make sure that the lod range is properly closed.

BUG:434647