Rename KisImage::lock() into KisImage::immediateLockForReadOnly()
If we lock for write we must regenerate LodN planes, which
is not possible right inside the stroke. Therefore, if you
need to lock the image for write you need to use
image->barrierLock(). It ensures that all the strokes are
finished before you do any modifications to the image.

CCBUG:434647