Add "allowOverwrite" to renameTag() in KisTagModel
Before this commit, KisTagModel wouldn't recognize
a situation where a tag was renamed to have an url
that already exists in the database so cannot be used.
This commit ensures that the model does recognize it
and behaves accordingly to the allowOverwrite option:
- if it's false and there is already a tag with that url,
it will return false.
- if it's true and there is already a tag with that url,
it will untag all resources from that existing tag,
then remove that tag completely from the database, and then
proceed with renaming of the current tag.